% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BLUP.pheno}
\alias{BLUP.pheno}
\title{Phenotype}
\format{
A numeric vector of grain yield values which were adjusted from the variability of the different locations and years.
}
\usage{
BLUP.pheno
}
\description{
Grain yield (Mg/ha)  of samples. This data was published by Guo et al. (2019) <doi:10.1016/j.molp.2018.12.022>
}
\examples{
data(BLUP.pheno)
}
\keyword{datasets}
