% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinship.R
\name{kinship}
\alias{kinship}
\title{Relationship matrix}
\usage{
kinship(X)
}
\arguments{
\item{X}{A numerical matrix of genotype. If is additive relationship matrix can be coded as -1,0,and 1.If is dominance relationship matrix can be coded as 0,1,and 0 (row: sample; column: marker).}
}
\value{
K A numerical matrix (relationship matrix)
}
\description{
This function calculate additive and dominance relationship matrix.
}
\examples{
KA <- kinship(train.geno)
}
