% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MO.pheno}
\alias{MO.pheno}
\title{Phenotype}
\format{
A numeric vector of grain yield values at Columbia, MO.
}
\usage{
MO.pheno
}
\description{
Grain yield (Mg/ha)  of samples at Columbia, MO. This data was published by Guo et al. (2019) <doi:10.1016/j.molp.2018.12.022>
}
\examples{
data(MO.pheno)
}
\keyword{datasets}
