% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parent.geno}
\alias{parent.geno}
\title{Genotype information of parental lines}
\format{
A numeric matrix with 24 rows (parental lines) and 46,134 columns (SNPs).
}
\usage{
parent.geno
}
\description{
Genotype information of parental lines. This data was published by Guo et al. (2019) <doi:10.1016/j.molp.2018.12.022>
}
\examples{
data(parent.geno)
}
\keyword{datasets}
