% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{train.geno}
\alias{train.geno}
\title{Genotype information of training set.}
\format{
A numeric matrix with 276 rows (F1s of half diallel design) and 46,134 columns (SNPs).
}
\usage{
train.geno
}
\description{
Genotype information of F1s used to construct a prediction model. This data was published by Guo et al. (2019) <doi:10.1016/j.molp.2018.12.022>
}
\examples{
data(train.geno)
}
\keyword{datasets}
