% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{train.pheno}
\alias{train.pheno}
\title{Phenotype}
\format{
A numeric data frame with four columns representing the hybrids, parent names, and their corresponding phenotypes.
}
\usage{
train.pheno
}
\description{
Fruit weight (kg) of samples. This data was published by Wu et al. (2019) <doi.org/10.3835/plantgenome2018.10.0082>
}
\examples{
data(train.pheno)
}
\keyword{datasets}
