% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGS.R
\name{BGS}
\alias{BGS}
\title{The Bayesian Gibbs sampling algorithm}
\usage{
BGS(
  pheno,
  KA,
  KD,
  mu.ini = 0,
  ga.ini = 0,
  gd.ini = 0,
  vE.ini = 1,
  vA.ini = 0,
  vD.ini = 0,
  iter = 5000,
  m = 1,
  S_star,
  nu_star = 5,
  cat.itr = TRUE
)
}
\arguments{
\item{pheno}{A numeric vector of phenotype which include all the lines in relationship matrix.}

\item{KA}{Additive relationship matrix.}

\item{KD}{Dominance relationship matrix.}

\item{mu.ini}{Initial values for the constant term.}

\item{ga.ini}{Initial values for the additive genotypic value.}

\item{gd.ini}{Initial values for the dominance genotypic value}

\item{vE.ini}{Initial values for the error variance.}

\item{vA.ini}{Initial values for the additive variance.}

\item{vD.ini}{Initial values for the dominance variance.}

\item{iter}{Number of iteration in each chain.}

\item{m}{Number of independent chain.}

\item{S_star}{The shape of freedom of the variance components for the prior distribution of a scaled inverse Chi square.}

\item{nu_star}{The degree of freedom of the variance components for the prior distribution of a scaled inverse Chi square.}

\item{cat.itr}{A boolean variable determines whether to print out the iteration number.}
}
\value{
This function will return the MCMC value of additive genotypic value (g_A;ga), dominance genotypic value (g_D;gd), constant term (\eqn{mu};mu), and variance components(additive variance;vA, dominance variance;vD, error variance;vE)
}
\description{
The Bayesian Gibbs sampling algorithm
}
\examples{

KA <- kinship(train.geno)
Xd <- replace(train.geno,train.geno==0,99)
Xd <- replace(Xd,Xd==1,0)
Xd <- replace(Xd,Xd==-1,0)
Xd <- replace(Xd,Xd==99,1)

KD <- kinship(Xd)
KA <- KA +diag(10^(-9),nrow = nrow(KA), ncol = ncol(KA))
KD <- KD +diag(10^(-9),nrow = nrow(KD), ncol = ncol(KD))

result <- BGS(pheno =train.pheno$F1.weight,KA = KA,KD = KD,mu.ini = 0,ga.ini = 0,gd.ini = 0,
vE.ini = 1,vA.ini = 0.5,vD.ini = 0.5,iter = 100,m = 1,S_star = 0.5*var(train.pheno$F1.weight),nu_star = 5)
}
