% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EHPGS.R
\name{EHPGS}
\alias{EHPGS}
\title{Evaluation of hybrid performance in plant breeding via genomic selection}
\usage{
EHPGS(
  num.F1,
  num.P,
  train.pheno,
  train.geno,
  parent.geno,
  train.in.test = NULL,
  method = "BGS",
  mu.ini = NULL,
  ga.ini = 0,
  gd.ini = 0,
  vE.ini = 1,
  vA.ini = 0.5,
  vD.ini = 0.5,
  iter = 10000,
  m = 5,
  burnIn = 500,
  thin = 5
)
}
\arguments{
\item{num.F1}{An integer number of superior hybrid combinations are select.}

\item{num.P}{An integer number of potential parental lines are select.}

\item{train.pheno}{A numeric vector of training set's phenotype.}

\item{train.geno}{A numerical matrix of training set's genotype. Matrix is coded as -1,0 and 1 for Minor, Hetero, and Major, respectively.}

\item{parent.geno}{A numerical matrix of parental lines' genotype. Matrix is coded as -1,0 and 1 for Minor, Hetero, and Major, respectively.}

\item{train.in.test}{A character vector of lines' names which are both in training set and half-diallel design.}

\item{method}{Select the method used to build the model, including BGS, BGLR, and sommer.}

\item{mu.ini}{Initial values for the constant term for BGS method.}

\item{ga.ini}{Initial values for the additive genotypic value for BGS method.}

\item{gd.ini}{Initial values for the dominance genotypic value for BGS method.}

\item{vE.ini}{Initial values for the error variance for BGS method.}

\item{vA.ini}{Initial values for the additive variance for BGS method.}

\item{vD.ini}{Initial values for the dominance variance for BGS method.}

\item{iter}{Number of iteration in each chain for BGS or BGLR method.}

\item{m}{Number of independent chain for BGS method.}

\item{burnIn}{An integer number of burn-in for BGLR method.}

\item{thin}{An integer number of thinning for BGLR method.}
}
\value{
This function will return 3 data frames, F1.out includes the top n GEBV F1s (n=num.F1) and its GEBV,SCA,MPH, and BPH. P.out includes the top m GEBV parents (m=num.P) and its GEBV,GCA. V.out includes the estimate of constant term and variance components (additive variance,dominance variance,and error variance)
}
\description{
Evaluation of hybrid performance in plant breeding via genomic selection
}
\examples{
parent.geno = parent.geno[1:10,]
hybrid.name = combn(row.names(parent.geno), 2, paste, collapse = ":")
tr.test = intersect(row.names(train.geno),combn(row.names(parent.geno), 2, paste, collapse = ":"))
EHPGS(num.F1=15,num.P=5,train.pheno=train.pheno$F1.weight,
train.geno = train.geno,parent.geno = parent.geno,train.in.test = tr.test,method= 'BGS',
mu.ini=NULL,ga.ini = 0,gd.ini = 0,vE.ini = 1,vA.ini = 0.5,vD.ini = 0.5,iter = 100,m = 1)

}
