% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parent.geno}
\alias{parent.geno}
\title{Genotype information of parental lines}
\format{
A numeric matrix with 142 rows (parental lines) and 4,521 columns (SNPs).
}
\usage{
parent.geno
}
\description{
Genotype information of parental lines. This data was published by Wu et al. (2019) <doi.org/10.3835/plantgenome2018.10.0082>
}
\examples{
data(parent.geno)
}
\keyword{datasets}
