% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{train.geno}
\alias{train.geno}
\title{Genotype information of training set.}
\format{
A numeric matrix with 119 rows (Hybrid of half diallel design) and 4,512 columns (SNPs).
}
\usage{
train.geno
}
\description{
Genotype information of hybrids used to construct a prediction model. This data was published by Wu et al. (2019) <doi.org/10.3835/plantgenome2018.10.0082>
}
\examples{
data(train.geno)
}
\keyword{datasets}
